Fix loading references if mapped by uuid and not by id.
DocumentManager::getReference() does not work correctly if called with an uuid instead of an id,
so if we have an uuid here, we have to fetch the node first (which unfortunately is not optimal
from a performance point of view) to get the correct id of the document.
(DocumentManager::getReference() seems to work first - but it will initialize the id field with the
uuid and this will cause errors later, e.g. when accessing other referenced documents of the loaded
document - it is simply not possible to use it correctly with an uuid)